pass command line test
